+++ physics-fe/src/components/Reg.vue
@@ -16,7 +16,7 @@
 <script>
 import { mapActions } from "vuex";
 export default {
-  name: "reg",
+  name: "Reg",
   data() {
     return { login: "", password: "" };
   },
