--- physics-fe/src/views/Home.vue
@@ -1,22 +1,0 @@
-<template>
-  <div>
-    <HelloWorld v-bind:user="user" />
-  </div>
-</template>
-
-<script>
-import HelloWorld from "@/components/HelloWorld.vue";
-
-export default {
-  name: "Home",
-  data() {
-    return {
-      user: "admin",
-      
-    };
-  },
-  components: {
-    HelloWorld,
-  },
-};
-</script>
